+++ CHANGELOG.md
@@ -1,7 +1,12 @@
-## 0.1.0
+## 0.2.0
 
-* Initial development release.
+* Adds support for CH34x-based devices.
+* Adds support for manual driver selection.
 
 ## 0.1.1
 
 * Fixes dependency resolution.
+
+## 0.1.0
+
+* Initial development release.